+++ NEWS.md
@@ -15,6 +15,7 @@
 * Quote query table and function names to support mixed case
 * Fix `collection` reponse to include actual extent (#54)
 * Fix `collection` response JSON `extent` property to be OAPIF-compliant (#54)
+* Fix landing page document to include links to resources `service-desc` and `conformance`
 
 ## Version 1.2
 *Released: 2020 Dec 1*